refactor: User entity에 enums 패키지 추가
