+++ src/main/java/com/project/dugoga/domain/user/entity/enums/UserRoleEnum.java
@@ -1,4 +1,4 @@
-package com.project.dugoga.domain.user.entity;
+package com.project.dugoga.domain.user.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
